Tagged content, separation of content and code, configurable primary menu.
Implemented tagging of content and tag summary pages.

Moved as much of the configuration into ./content as possible.

Excluded ./content from this repo so that content can be tracked in a
separate repo.

Implemented configurable primary menu so that pages may or may not have
primary menu associated with them. Additionally, provided a means by which
categories and author name can be turned on or off depending on the needs
of the user.

Added a script to bootstrap the site and optionally, to copy starter content.

Added a script to generate new posts and pages.