Added ./content/.keep-content-dir to this repo.
To separate content from code, ./content is excluded in .gitignore.
Git will not keep empty directories so ./content/.keep-content-dir
was added to force ./content to remain in the source tree for this
repo while ignoring the contents of that directory except for that
one file. That directory contains user-generated content and site
customizations, which can now be kept under revision control in a
separate repository.